Deploy baremetal bootstrap services with MCO
Currently we have two very different sets of manifests and templates
for deploying the baremetal infrastructure services. One lives in the
installer, the other lives in MCO. This change removes the services
from the installer and makes use of the files being added to the MCO
that serve the same purpose but keep all of the templates in the same
repo and allow us to use the same MCO data for populating dynamic
values.